# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ConnectWalletRequest {
  walletAddress: string;
}

export interface User {
  id: string;
  walletAddress: string;
  createdAt: string;
}

export interface Wallet {
  id: string;
  userId: string;
  balance: number;
  totalEarned: number;
  totalSpent: number;
  stakedAmount: number;
  stakingRewards: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConnectWalletResponse {
  user: User;
  wallet: Wallet;
  isNew: boolean;
}

export interface SessionResponse {
  user?: User;
  wallet?: Wallet;
}

export interface Fighter {
  id: string;
  tokenId: string;
  name: string;
  description?: string | null;
  imageUrl?: string | null;
  ownerId: string;
  aggression: number;
  defense: number;
  speed: number;
  power: number;
  intelligence: number;
  level: number;
  experience: number;
  wins: number;
  losses: number;
  totalBattles: number;
  trainingCount: number;
  lastTrainedAt?: string | null;
  color: string;
  skinId?: string | null;
  txHash?: string | null;
  isMinted: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface FighterListResponse {
  fighters: Fighter[];
}

export interface MintFighterRequest {
  name: string;
  description?: string | null;
  color?: string | null;
}

export interface BattleRound {
  round: number;
  attacker: string;
  defender: string;
  damage: number;
  fighter1Hp: number;
  fighter2Hp: number;
  move: string;
}

export type BattleStatus = (typeof BattleStatus)[keyof typeof BattleStatus];

export const BattleStatus = {
  PENDING: "PENDING",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
} as const;

export interface Battle {
  id: string;
  fighter1Id: string;
  fighter2Id: string;
  fighter1: Fighter;
  fighter2: Fighter;
  winnerId?: string | null;
  loserId?: string | null;
  rounds: number;
  fighter1FinalHp: number;
  fighter2FinalHp: number;
  battleLog?: BattleRound[] | null;
  prizePool: number;
  winnerReward: number;
  status: BattleStatus;
  startedAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
}

export interface BattleListResponse {
  battles: Battle[];
}

export interface CreateBattleRequest {
  fighter1Id: string;
  fighter2Id: string;
}

export type TrainFighterRequestType =
  (typeof TrainFighterRequestType)[keyof typeof TrainFighterRequestType];

export const TrainFighterRequestType = {
  BASIC: "BASIC",
  ADVANCED: "ADVANCED",
  INTENSIVE: "INTENSIVE",
  AI_OPTIMIZED: "AI_OPTIMIZED",
} as const;

export type TrainFighterRequestStat =
  (typeof TrainFighterRequestStat)[keyof typeof TrainFighterRequestStat];

export const TrainFighterRequestStat = {
  aggression: "aggression",
  defense: "defense",
  speed: "speed",
  power: "power",
  intelligence: "intelligence",
} as const;

export interface TrainFighterRequest {
  type: TrainFighterRequestType;
  stat: TrainFighterRequestStat;
}

export interface TrainFighterResponse {
  fighter: Fighter;
  improvement: number;
  cost: number;
  statImproved: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  MINT_FIGHTER: "MINT_FIGHTER",
  TRAIN_FIGHTER: "TRAIN_FIGHTER",
  BATTLE_REWARD: "BATTLE_REWARD",
  BATTLE_ENTRY: "BATTLE_ENTRY",
  STAKE: "STAKE",
  UNSTAKE: "UNSTAKE",
  STAKING_REWARD: "STAKING_REWARD",
} as const;

export interface Transaction {
  id: string;
  walletId: string;
  type: TransactionType;
  amount: number;
  description?: string | null;
  battleId?: string | null;
  fighterId?: string | null;
  createdAt: string;
}

export interface TransactionListResponse {
  transactions: Transaction[];
}

export interface StakeRequest {
  amount: number;
}

export interface PlatformStats {
  totalFighters: number;
  totalBattles: number;
  totalUsers: number;
  totalOneDistributed: number;
  topFighters: Fighter[];
}