# lib/api-zod/src/generated/types/battleRound.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */

export interface BattleRound {
  round: number;
  attacker: string;
  defender: string;
  damage: number;
  fighter1Hp: number;
  fighter2Hp: number;
  move: string;
}