# lib/api-zod/src/generated/types/connectWalletResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";
import type { Wallet } from "./wallet";

export interface ConnectWalletResponse {
  user: User;
  wallet: Wallet;
  isNew: boolean;
}