# lib/api-zod/src/generated/types/mintFighterRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */

export interface MintFighterRequest {
  name: string;
  description?: string | null;
  color?: string | null;
}