# lib/api-zod/src/generated/types/trainFighterRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */
import type { TrainFighterRequestStat } from "./trainFighterRequestStat";
import type { TrainFighterRequestType } from "./trainFighterRequestType";

export interface TrainFighterRequest {
  type: TrainFighterRequestType;
  stat: TrainFighterRequestStat;
}