# lib/api-zod/src/generated/types/trainFighterResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */
import type { Fighter } from "./fighter";

export interface TrainFighterResponse {
  fighter: Fighter;
  improvement: number;
  cost: number;
  statImproved: string;
}