# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arena GameFi API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  walletAddress: string;
  createdAt: Date;
}